make AD default and dispersion in metrics
